requirements.txt
pytorch
torchvision
pytest
tqdm
matplotlib
imageio
